requirements.txt
# Water Quality Scraper - Core Dependencies
playwright>=1.41
SQLAlchemy>=1.4
PyYAML>=6.0
python-dateutil>=2.8
pytz>=2023.3
pydantic>=1.10,<3.0
